+++ src/Framework/OutboundAdapters/OutboundAbstractAdapter.php
@@ -29,27 +29,24 @@ class OutboundAbstractAdapter implements BrokerOutboundAdapterInterface
     protected string $replyTo;
     protected bool $isSyncOverAsync = false;
 
-    public function __construct()
-    {
-        $this->application = app();
-    }
-
     /**
      * Nobody cares about the implementation
      *
      * @param MessageBagInterface $message
+     * @param Application $application
      *
      * @return BrokerResponse|null
      */
-    public function __invoke(MessageBagInterface $message): ?BrokerResponse
+    public function __invoke(MessageBagInterface $message, Application $application): ?BrokerResponse
     {
-        return $this->send($message);
+        $this->application = $application;
+        return $this->push($message);
     }
 
     /**
      * @inheritdoc
      */
-    public function send(MessageBagInterface $message, int $timeout = 30): ?BrokerResponse
+    public function push(MessageBagInterface $message, int $timeout = 30): ?BrokerResponse
     {
         if ($this->isSyncOverAsync) {
             $this->channelName = "";
@@ -71,7 +68,7 @@ class OutboundAbstractAdapter implements BrokerOutboundAdapterInterface
         $publisher->publish($message, $this->channelName);
 
         if ($this->isSyncOverAsync) {
-            return $this->pull($timeout);
+            return $this->pull($timeout, $message);
         }
 
         return null;
@@ -80,7 +77,7 @@ class OutboundAbstractAdapter implements BrokerOutboundAdapterInterface
     /**
      * @inheritdoc
      */
-    public function pull(int $timeout = 30): ?BrokerResponse
+    public function pull(int $timeout = 30, MessageBagInterface $context = null): ?BrokerResponse
     {
         /** @var AMQPChannel $channel */
         $channel = ($this->application->get("brokerStreamConnection"))->channel();
@@ -90,7 +87,11 @@ class OutboundAbstractAdapter implements BrokerOutboundAdapterInterface
             do {
                 $response = $channel->basic_get($this->replyTo);
                 if (!is_null($response)) {
-                    break;
+                    if ($this->isResponseForContext($response, $context)) {
+                        break;
+                    }
+                    // remove this message from the queue
+                    $response->nack();
                 }
                 // wait a while - prevent CPU load
                 usleep(5000);
@@ -133,5 +134,26 @@ class OutboundAbstractAdapter implements BrokerOutboundAdapterInterface
     {
         return (new InfrastructureStreamer($this->application))
             ->brokerActiveRpcSetup();
+    }
+
+    /**
+     * @param AMQPMessage $response
+     * @param MessageBagInterface|null $context
+     *
+     * @return bool
+     */
+    protected function isResponseForContext(AMQPMessage $response, MessageBagInterface $context = null): bool
+    {
+        // if no context is provided, this will always return true
+        if (is_null($context)) {
+            return true;
+        }
+
+        $correlationId = $response->get_properties()["correlation_id"] ?? null;
+        if (!is_null($correlationId) && $context->getProperty("correlation_id") === $correlationId) {
+            return true;
+        }
+
+        return false;
     }
 }